rbdd/requirements.txt
numpy
pandas
matplotlib
seaborn
pycryptodome
loguru
pytest
scipy
tabulate
